+++ internal/clients/git/git.go
@@ -293,9 +293,9 @@ func (r *Repo) MakeTrackingBranches(remoteName string) error {
 	return nil
 }
 
-func (r *Repo) FetchAll() error {
+func (r *Repo) FetchAll(indent int) error {
 	if err := r.repository.Fetch(&git.FetchOptions{
-		Progress: os.Stdout,
+		Progress: newIndentedWriter(indent, os.Stdout),
 		Tags:     git.AllTags,
 		RefSpecs: []config.RefSpec{
 			"+refs/heads/*:refs/heads/*",
@@ -436,7 +436,7 @@ func Open(local string) (*Repo, error) {
 	}, errors.Wrapf(err, "couldn't open git repo at %s", local)
 }
 
-func Clone(remote, local string) (*Repo, error) {
+func Clone(indent int, remote, local string) (*Repo, error) {
 	u, err := url.Parse(remote)
 	if err != nil {
 		return nil, errors.Wrapf(err, "couldn't parse %s as a url", remote)
@@ -447,14 +447,14 @@ func Clone(remote, local string) (*Repo, error) {
 	remote = u.String()
 	repo, err := git.PlainClone(local, false, &git.CloneOptions{
 		URL:      remote,
-		Progress: os.Stdout,
+		Progress: newIndentedWriter(indent, os.Stdout),
 	})
 	return &Repo{
 		repository: repo,
 	}, errors.Wrapf(err, "couldn't clone git repo %s to %s", remote, local)
 }
 
-func CloneMirrored(remote, local string) (*Repo, error) {
+func CloneMirrored(indent int, remote, local string) (*Repo, error) {
 	u, err := url.Parse(remote)
 	if err != nil {
 		return nil, errors.Wrapf(err, "couldn't parse %s as a url", remote)
@@ -465,7 +465,7 @@ func CloneMirrored(remote, local string) (*Repo, error) {
 	remote = u.String()
 	repo, err := git.PlainClone(local, false, &git.CloneOptions{
 		URL:      remote,
-		Progress: os.Stdout,
+		Progress: newIndentedWriter(indent, os.Stdout),
 		Mirror:   true,
 	})
 	return &Repo{
@@ -499,13 +499,13 @@ func Prune(local string) (updated bool, err error) {
 	return true, nil
 }
 
-func Fetch(local string) (updated bool, err error) {
+func Fetch(indent int, local string) (updated bool, err error) {
 	repo, err := git.PlainOpen(local)
 	if err != nil {
 		return false, errors.Wrapf(err, "couldn't open %s as git repo", local)
 	}
 	if err = repo.Fetch(&git.FetchOptions{
-		Progress: os.Stdout,
+		Progress: newIndentedWriter(indent, os.Stdout),
 		Tags:     git.AllTags,
 		RefSpecs: []config.RefSpec{
 			"+refs/heads/*:refs/remotes/origin/*",
@@ -519,7 +519,7 @@ func Fetch(local string) (updated bool, err error) {
 	return true, nil
 }
 
-func Pull(local string) (updated bool, err error) {
+func Pull(indent int, local string) (updated bool, err error) {
 	repo, err := git.PlainOpen(local)
 	if err != nil {
 		return false, errors.Wrapf(err, "couldn't open %s as git repo", local)
@@ -529,7 +529,7 @@ func Pull(local string) (updated bool, err error) {
 		return false, err
 	}
 	if err = worktree.Pull(&git.PullOptions{
-		Progress: os.Stdout,
+		Progress: newIndentedWriter(indent, os.Stdout),
 	}); err != nil {
 		if errors.Is(err, git.NoErrAlreadyUpToDate) {
 			return false, nil